Move device dependent code to amdgpu.c and radeon.c
